+++ src/app/components/AccountTable/AccountTable.tsx
@@ -1,8 +1,8 @@
 'use client'
-import TransactionRow from '../TransactionRow'
+import TransactionRow from './TransactionRow'
 import { Bookmark } from 'lucide-react'
 import ResizableColumn from '../ResizableColumn'
-import TransactionForm from '../TransactionForm'
+import TransactionForm from './TransactionForm'
 import { useState, useEffect, useMemo } from 'react'
 import AccountsHeader from './AccountsHeader'
 import ActionBar from './ActionBar'
